Added Tutorial Document
Added the initial release of the tutorial document on how to build and configure the RasPi for VeinCam. Currently only has the configuration/preparation steps, and more will be added as software becomes more complete, instead of test harnesses.

Updated the meeting minutes for the class tutorial with my input regarding the tutorial document.